updated deps and renamed projs
