+++ CMakeLists.txt
@@ -92,7 +92,7 @@ if(MSVC)
 endif()
 
 # set the startup project for the "play" button in MSVC
-set_property(DIRECTORY PROPERTY VS_STARTUP_PROJECT intro)
+set_property(DIRECTORY PROPERTY VS_STARTUP_PROJECT tagdetect)
 
 if(CMAKE_SKIP_INSTALL_RULES)
   return()
@@ -104,7 +104,7 @@ include(cmake/PackageProject.cmake)
 # we know we want to ship
 apriltagexperiments_package_project(
   TARGETS
-  intro
+  tagdetect
   apriltagexperiments_options
   apriltagexperiments_warnings
   # FIXME: this does not work! CK
