--- include/apriltagexperiments/sample_library.hpp
@@ -1,15 +1,0 @@
-#ifndef SAMPLE_LIBRARY_HPP
-#define SAMPLE_LIBRARY_HPP
-
-#include <apriltagexperiments/sample_library_export.hpp>
-
-[[nodiscard]] SAMPLE_LIBRARY_EXPORT int factorial(int) noexcept;
-
-[[nodiscard]] constexpr int factorial_constexpr(int input) noexcept
-{
-  if (input == 0) { return 1; }
-
-  return input * factorial_constexpr(input - 1);
-}
-
-#endif
