+++ src/tagdetect/CMakeLists.txt
@@ -1,19 +1,16 @@
-add_executable(intro main.cpp)
+add_executable(tagdetect main.cpp)
 
 target_link_libraries(
-  intro
+  tagdetect
   PRIVATE apriltagexperiments::apriltagexperiments_options
           apriltagexperiments::apriltagexperiments_warnings)
 
 target_link_system_libraries(
-  intro
+  tagdetect
   PRIVATE
           CLI11::CLI11
           fmt::fmt
           spdlog::spdlog
-          lefticus::tools
-          ftxui::screen
-          ftxui::dom
-          ftxui::component)
+          lefticus::tools)
 
-target_include_directories(intro PRIVATE "${CMAKE_BINARY_DIR}/configured_files/include")
+target_include_directories(tagdetect PRIVATE "${CMAKE_BINARY_DIR}/configured_files/include")
