+++ test/constexpr_tests.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <apriltagexperiments/sample_library.hpp>
+#include <apriltagexperiments/imgproc.hpp>
 
 TEST_CASE("Factorials are computed with constexpr", "[factorial]")
 {
