+++ test/tests.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 
-#include <apriltagexperiments/sample_library.hpp>
+#include <apriltagexperiments/imgproc.hpp>
 
 
 TEST_CASE("Factorials are computed", "[factorial]")
